Fixed bug in random generators that caused application fail with run-time error when being run for a long time. The problem was that random generators rarely returned upper bound of the input range.

+++ Raytracer/Math/SamplingRoutines.h
@@ -377,7 +377,7 @@ namespace SamplingRoutines
     else
       for (size_t i = 0; i < i_samples_num; ++i)
         {
-        unsigned int other = RandomInt(i_samples_num);
+        unsigned int other = (unsigned int) RandomInt((int)i_samples_num);
         std::swap(*(i_begin+i), *(i_begin+other));
         }
     }
